+++ next.config.js
@@ -1,3 +1,7 @@
 module.exports = {
-  webpack: config => Object.assign(config, { node: { fs: 'empty' } }),
+  webpack: config => Object.assign(config, {
+    node: {
+      fs: 'empty', child_process: 'empty', net: 'empty', tls: 'empty',
+    },
+  }),
 };